requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
pydantic==2.9.2
python-jose==3.3.0
